+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Set fetcher timeout to 3 minutes.
+- Set deliverer timeout to 30 seconds.
 
 ## [1.14.0] - 2023-02-22